requirements.txt
fastapi
uvicorn[standard]
httpx
beautifulsoup4
lxml
spacy
sentence-transformers
torch
faiss-cpu
numpy
pydantic
pydantic-settings
langdetect
dateparser
transformers
pymongo
motor
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
streamlit
plotly
pandas
python-multipart
playwright
pdfplumber
cryptography
email-validator
# --- Production Dependencies ---
slowapi
redis
celery[redis]
google-genai

